requirements.txt
Click==7.0
dominate==2.3.4
Flask==1.0.2
Flask-Bootstrap4==4.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
visitor==0.1.3
Werkzeug==0.14.1
picamera==1.13
Pillow==5.3.0
gpiozero==1.4.1
